Finished changing all possible ints to their uXX/sXX variants

+++ include/spritetools_render.h
@@ -17,10 +17,10 @@
 \*****************************/
 /* Inits rendering */
 /* Returns 1 on success and 0 on failure */
-int ST_RenderInit(void);
+u8 ST_RenderInit(void);
 
 /* Finishes rendering */
-int ST_RenderFini(void);
+u8 ST_RenderFini(void);
 
 /* Start frame */
 /* Takes screen (GFX_TOP or GFX_BOTTOM) */
@@ -46,7 +46,7 @@ u32 ST_RenderGetBackground(void);
 \*******************************/
 /* Draw Spritesheet at Position */
 /* Takes spritesheet and x and y of position to render on screen */
-void ST_RenderSpritesheetPosition(st_spritesheet *spritesheet, int x, int y);
+void ST_RenderSpritesheetPosition(st_spritesheet *spritesheet, s64 x, s64 y);
 
 /* Draw Spritesheet at 0,0 */
 /* Takes spritesheet */
@@ -58,17 +58,17 @@ void ST_RenderSpritesheet(st_spritesheet *spritesheet);
 /*   Takes width and height of the sprite in the spritesheet */
 /*   Takes position to print the sprite on screen */
 void ST_RenderSpritePosition(st_spritesheet *spritesheet,
-  unsigned int xleft, unsigned int ytop,
-  unsigned int width, unsigned int height,
-  int x, int y);
+  u32 xleft, u32 ytop,
+  u32 width, u32 height,
+  s64 x, s64 y);
 
 /* Draw Sprite in Spritesheet at 0,0 */
 /* Takes spritesheet */
 /*   Takes x and y of the top left pixel of the sprite in the spritesheet */
 /*   Takes width and height of the sprite in the spritesheet */
 void ST_RenderSprite(st_spritesheet *spritesheet,
-  unsigned int xleft, unsigned int ytop,
-  unsigned int width, unsigned int height);
+  u32 xleft, u32 ytop,
+  u32 width, u32 height);
 
 /* Draw Scaled Sprite in Spritesheet at Position */
 /* Takes spritesheet */
@@ -77,9 +77,9 @@ void ST_RenderSprite(st_spritesheet *spritesheet,
 /*   Takes position to print the sprite on screen */
 /*   Takes value to scale by */
 void ST_RenderSpritePositionScale(st_spritesheet *spritesheet,
-  unsigned int xleft, unsigned int ytop,
-  unsigned int width, unsigned int height,
-  int x, int y,
+  u32 xleft, u32 ytop,
+  u32 width, u32 height,
+  s64 x, s64 y,
   double scale);
 
 /* Draw Rotated Sprite in Spritesheet at Position */
@@ -89,9 +89,9 @@ void ST_RenderSpritePositionScale(st_spritesheet *spritesheet,
 /*   Takes position to print the sprite on screen */
 /*   Takes radian value to rotate by */
 void ST_RenderSpritePositionRotate(st_spritesheet *spritesheet,
-  unsigned int xleft, unsigned int ytop,
-  unsigned int width, unsigned int height,
-  int x, int y,
+  u32 xleft, u32 ytop,
+  u32 width, u32 height,
+  s64 x, s64 y,
   double rotate);
 
 /* Draw Scaled and Rotated Sprite in Spritesheet at Position */
@@ -102,9 +102,9 @@ void ST_RenderSpritePositionRotate(st_spritesheet *spritesheet,
 /*   Takes value to scale by */
 /*   Takes value to rotate by */
 void ST_RenderSpritePositionScaleRotate(st_spritesheet *spritesheet,
-  unsigned int xleft, unsigned int ytop,
-  unsigned int width, unsigned int height,
-  int x, int y,
+  u32 xleft, u32 ytop,
+  u32 width, u32 height,
+  s64 x, s64 y,
   double scale,
   double rotate);
 
@@ -117,9 +117,9 @@ void ST_RenderSpritePositionScaleRotate(st_spritesheet *spritesheet,
 /*   Takes value to rotate by */
 /*   Takes color to blend with (rgba8) */
 void ST_RenderSpriteAdvanced(st_spritesheet *spritesheet,
-  unsigned int xleft, unsigned int ytop,
-  unsigned int width, unsigned int height,
-  int x, int y,
+  u32 xleft, u32 ytop,
+  u32 width, u32 height,
+  s64 x, s64 y,
   double scale,
   double rotate,
   u8 red, u8 green, u8 blue, u8 alpha);
@@ -129,21 +129,21 @@ void ST_RenderSpriteAdvanced(st_spritesheet *spritesheet,
 \*************************/
 /* Draw frame at given position */
 /* Takes spritesheet and position at which to draw */
-void ST_RenderFramePosition(st_frame *frame, int x, int y);
+void ST_RenderFramePosition(st_frame *frame, s64 x, s64 y);
 
 /* Draw scaled frame at given position */
 /* Takes spritesheet, position at which to draw and a scalar multiplier */
-void ST_RenderFrameScale(st_frame *frame, int x, int y, double scale);
+void ST_RenderFrameScale(st_frame *frame, s64 x, s64 y, double scale);
 
 /* Draw rotated frame at given position */
 /* Takes spritesheet, position at which to draw and a rotation in radians */
-void ST_RenderFrameRotate(st_frame *frame, int x, int y, double rotate);
+void ST_RenderFrameRotate(st_frame *frame, s64 x, s64 y, double rotate);
 
 /* Draw scaled, rotated, and blended frame at given position */
 /* Takes spritesheet and position at which to draw */
 /*   Takes scalar multiplier and rotation in radians */
 /*   Takes red, green, blue, and alpha of color to blend */
-void ST_RenderFramePositionAdvanced(st_frame *frame, int x, int y,
+void ST_RenderFramePositionAdvanced(st_frame *frame, s64 x, s64 y,
   double scale, double rotate,
   u8 red, u8 green, u8 blue, u8 alpha);
 
@@ -152,22 +152,22 @@ void ST_RenderFramePositionAdvanced(st_frame *frame, int x, int y,
 \*****************************/
 /* Draw the current frame of an animation at given position */
 /* Takes a pointer to an animation and a position */
-void ST_RenderAnimationCurrent(st_animation *animation, int x, int y);
+void ST_RenderAnimationCurrent(st_animation *animation, s64 x, s64 y);
 
 /* Draw the next frame of an animation at given position */
 /*   Also adds 1 to the current frame */
 /* Takes a pointer to an animation and a position */
-void ST_RenderAnimationNext(st_animation *animation, int x, int y);
+void ST_RenderAnimationNext(st_animation *animation, s64 x, s64 y);
 
 /* Draw the previous frame of an animation at given position */
 /*   Also subtracts 1 from the current frame */
 /* Takes a pointer to an animation and a position */
-void ST_RenderAnimationPrevious(st_animation *animation, int x, int y);
+void ST_RenderAnimationPrevious(st_animation *animation, s64 x, s64 y);
 
 /* Plays an animation at given position */
 /*   This also accounts for the animation's speed */
 /* Takes a pointer to an animation and a position */
-void ST_RenderAnimationPlay(st_animation *animation, int x, int y);
+void ST_RenderAnimationPlay(st_animation *animation, s64 x, s64 y);
 
 /****************************************\
 |*     Advanced Animation Rendering     *|
@@ -176,19 +176,19 @@ void ST_RenderAnimationPlay(st_animation *animation, int x, int y);
 /*   rotation in radians, and red, green, blue, and alpha of a */
 /*   color to blend with */
 
-void ST_RenderAnimationCurrentAdvanced(st_animation *animation, int x, int y,
+void ST_RenderAnimationCurrentAdvanced(st_animation *animation, s64 x, s64 y,
   double scale, double rotate,
   u8 red, u8 green, u8 blue, u8 alpha);
 
-void ST_RenderAnimationNextAdvanced(st_animation *animation, int x, int y,
+void ST_RenderAnimationNextAdvanced(st_animation *animation, s64 x, s64 y,
   double scale, double rotate,
   u8 red, u8 green, u8 blue, u8 alpha);
 
-void ST_RenderAnimationPreviousAdvanced(st_animation *animation, int x, int y,
+void ST_RenderAnimationPreviousAdvanced(st_animation *animation, s64 x, s64 y,
   double scale, double rotate,
   u8 red, u8 green, u8 blue, u8 alpha);
 
-void ST_RenderAnimationPlayAdvanced(st_animation *animation, int x, int y,
+void ST_RenderAnimationPlayAdvanced(st_animation *animation, s64 x, s64 y,
   double scale, double rotate,
   u8 red, u8 green, u8 blue, u8 alpha);
 
@@ -198,6 +198,6 @@ void ST_RenderAnimationPlayAdvanced(st_animation *animation, int x, int y,
 /* Plays the current animation of an entity by name */
 /* Takes a pointer to an entity */
 /* Returns 1 on success and 0 on failure */
-int ST_RenderEntity(st_entity *entity);
+u8 ST_RenderEntity(st_entity *entity);
 
 #endif
